Move decoding to sample project
