+++ zano-ios-sample/JSONRPC/Commons/AnyCodableError.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public enum AnyCodableError: Error {
+    case nullFound
+}
