+++ zano-ios-sample/JSONRPC/JSONRPCImports.swift
@@ -1,0 +1,3 @@
+//#if !CocoaPods
+//@_exported import Commons
+//#endif
