--- zano-ios/JSONRPC/Commons/AnyCodableError.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public enum AnyCodableError: Error {
-    case nullFound
-}
