--- zano-ios/JSONRPC/JSONRPCImports.swift
@@ -1,3 +1,0 @@
-//#if !CocoaPods
-//@_exported import Commons
-//#endif
