+++ zano-ios/zano.swift
@@ -21,7 +21,7 @@ public var asyncCallMethods: [String] {
 
 public enum ZanoWallet {
     // MARK: - Initialization Functions
-    public static func InitIpPort(ip: String, port: String, working_dir: String, log_level:Int32) throws -> SuccessResult {
+    public static func InitIpPort(ip: String, port: String, working_dir: String, log_level:Int32) -> String {
         debugPrint("ip is \(ip)")
         debugPrint("port is \(port)")
         debugPrint("working_dir is \(working_dir)")
@@ -32,93 +32,68 @@ public enum ZanoWallet {
         debugPrint("ipCpp is \(ipCpp)")
         debugPrint("portCpp is \(portCpp)")
         debugPrint("workingDirRawCpp is \(workingDirRawCpp)")
+        
         let jsonStr = String(ZanoCore.initIpPort(ipCpp, portCpp, workingDirRawCpp, log_level))
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseResponse(jsonData: jsonData, resultType: SuccessResult.self)
-        return result
-        
+        return jsonStr
     }
     
     // MARK: - Utility Functions
-    public static func reset() throws -> SuccessResult {
+    //    close all opened wallets without saving files
+//    public static func reset() throws -> SuccessResult {
+    public static func reset() -> String {
+        debugPrint("\(#function) starts")
         
         let jsonStr = String(ZanoCore.reset())
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseResponse(jsonData: jsonData, resultType: SuccessResult.self)
-        return result
+        return jsonStr
     }
     
     public static func getVersion() -> String {
-        String(ZanoCore.get_version())
+        debugPrint("\(#function) starts")
+        return String(ZanoCore.get_version())
     }
-    
-    public static func getWalletFiles() throws -> GetWalletFileResult {
+
+    public static func getWalletNames() -> String {
+        debugPrint("\(#function) starts")
         let jsonStr = String(ZanoCore.get_wallet_files())
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseJsonResponse(jsonData: jsonData, resultType: GetWalletFileResult.self)
-        return result
+        return jsonStr
     }
     
-    public static func deleteWallet(file_name: String) throws -> SuccessResult {
+    public static func deleteWallet(file_name: String) -> String {
+        debugPrint("\(#function) starts")
         let fileName = StringConversionUtils.swiftStringToCppString(file_name)
         let jsonStr = String(ZanoCore.delete_wallet(fileName))
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseResponse(jsonData: jsonData, resultType: SuccessResult.self)
-        return result
+        return jsonStr
     }
-    
-    public static func getAddressInfo(addr: String) throws -> AddressInfo {
+
+    public static func getAddressInfo(addr: String) -> String {
+        debugPrint("\(#function) starts")
         let address = StringConversionUtils.swiftStringToCppString(addr)
         let jsonStr = String(ZanoCore.get_address_info(address))
-        
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseJsonResponse(jsonData: jsonData, resultType: AddressInfo.self)
-        return result
+        return jsonStr
     }
     
     // MARK: - Configuration Functions
     public static func getLogsBuffer() -> String {
-        String(ZanoCore.get_logs_buffer())
+        debugPrint("\(#function) starts")
+        return String(ZanoCore.get_logs_buffer())
     }
     
-    public static func truncateLog() throws -> SuccessResult {
+    public static func truncateLog() -> String {
+        debugPrint("\(#function) starts")
         let jsonStr = String(ZanoCore.truncate_log())
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
         
-        let result = try JSONRPCParser.parseResponse(jsonData: jsonData, resultType: SuccessResult.self)
-        return result
+        return jsonStr
     }
-    
-    public static func getConnectivityStatus() throws -> ConnectivityStatus {
+
+    public static func getConnectivityStatus() -> String {
+        debugPrint("\(#function) starts")
         let jsonStr = String(ZanoCore.get_connectivity_status())
-        
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseJsonResponse(jsonData: jsonData, resultType: ConnectivityStatus.self)
-        return result
+        return jsonStr
     }
     
     // MARK: - Wallet Management Functions
-    public static func restore(seed: String, walletName: String, password: String, seed_password:String) throws -> WalletResult {
+    public static func restore(seed: String, walletName: String, password: String, seed_password:String) -> String {
+        debugPrint("\(#function) starts")
         let sd = StringConversionUtils.swiftStringToCppString(seed)
         let pa = StringConversionUtils.swiftStringToCppString(walletName)
         let ps = StringConversionUtils.swiftStringToCppString(password)
@@ -130,37 +105,28 @@ public enum ZanoWallet {
         debugPrint("ps is \(ps)")
         debugPrint("sdps is \(sdps)")
         let jsonStr = String(ZanoCore.restore(sd, pa, ps, sdps))
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        let result = try JSONRPCParser.parseResponse(jsonData: jsonData, resultType: WalletResult.self)
-        return result
+        return jsonStr
     }
     
     // Create a new wallet
     // https://github.com/hyle-team/zano_mobile/blob/6cf344ee1d1b8f462c9339cce5054bc1fd2790b8/src/cpp/zano_wallet_interface_impl.cxx#L74
     // generate_new_wallet
-    public static func generate(walletName: String, password: String) throws -> WalletResult {
+    public static func generate(walletName: String, password: String) -> String {
+        debugPrint("\(#function) starts")
         let name = StringConversionUtils.swiftStringToCppString(walletName)
         let ps = StringConversionUtils.swiftStringToCppString(password)
-        let jsonStr = String(ZanoCore.generate(name, ps))
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
+        debugPrint("ps is \(name)")
+        debugPrint("sdps is \(ps)")
         
-        let result = try JSONRPCParser.parseResponse(jsonData: jsonData, resultType: WalletResult.self)
-        return result
+        let jsonStr = String(ZanoCore.generate(name, ps))
+        return jsonStr
     }
     
-    // not working?
-    public static func getOpenedWallets() throws -> [WalletResult] {
+    // response can be "{\r\n  \"id\": 0,\r\n  \"jsonrpc\": \"\"\r\n}"
+    public static func getOpenedWallets() -> String {
+        debugPrint("\(#function) starts")
         let jsonStr = String(ZanoCore.get_opened_wallets())
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseResponse(jsonData: jsonData, resultType: [WalletResult].self)
-        return result
+        return jsonStr
     }
     
     
@@ -175,8 +141,8 @@ public enum ZanoWallet {
      */
     
     // MARK: - Asynchronous API Functions
-    
-    public static func asyncCall(methodName: String, walletId: UInt64, params: String) async throws -> JobResult {
+    public static func asyncCall(methodName: String, walletId: UInt64, params: String) throws -> String {
+        debugPrint("\(#function) starts")
         guard asyncCallMethods.contains(methodName) else {
             throw ZANOError.unknown(message: "unsupported method name")
         }
@@ -184,75 +150,61 @@ public enum ZanoWallet {
         let name = StringConversionUtils.swiftStringToCppString(methodName)
         let param = StringConversionUtils.swiftStringToCppString(params)
         
+        debugPrint("------")
+        debugPrint(name)
+        debugPrint(walletId)
+        debugPrint(param)
         
         let jsonStr = String(ZanoCore.async_call(name, walletId, param))
-        
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseJsonResponse(jsonData: jsonData, resultType: JobResult.self)
-        return result
+        return jsonStr
     }
+    
     // MARK:  Wallet Operations
-    public static func asyncGetWalletStatus(jobId: UInt64) async throws -> WalletSyncStatus {
+    public static func asyncGetRestoreInfo(jobId: UInt64) async  -> String {
+        debugPrint("\(#function) starts")
         let jsonStr = String(String(ZanoCore.try_pull_result(jobId)))
-        
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseJsonResponse(jsonData: jsonData, resultType: WalletSyncStatus.self)
-        return result
+        return jsonStr
+    }
+    
+    
+    // MARK:  Wallet Operations
+    public static func asyncGetWalletStatus(jobId: UInt64) async -> String {
+        debugPrint("\(#function) starts")
+        let jsonStr = String(String(ZanoCore.try_pull_result(jobId)))
+        return jsonStr
     }
 
     
-    public static func asyncRestore(jobId: UInt64) async throws -> WalletResult {
+    public static func asyncRestore(jobId: UInt64) async -> String {
+        debugPrint("\(#function) starts")
         let jsonStr = String(String(ZanoCore.try_pull_result(jobId)))
-        
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseJsonResponse(jsonData: jsonData, resultType: WalletResult.self)
-        return result
+        return jsonStr
     }
     
     // https://docs.zano.org/docs/build/rpc-api/wallet-rpc-api/getbalance
-    public static func asyncGetBalance(jobId: UInt64) async throws -> BalanceResult {
+    public static func asyncGetBalance(jobId: UInt64) async -> String {
+        debugPrint("\(#function) starts")
         let jsonStr = String(String(ZanoCore.try_pull_result(jobId)))
-        
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseResponse(jsonData: jsonData, resultType: BalanceResult.self)
-        return result
+        return jsonStr
     }
 
     
     // required to get JobId from asyncCall
-    public static func asyncOpen(jobId: UInt64) async throws -> WalletResult {
-        let jsonStr = String(String(ZanoCore.try_pull_result(jobId)))
+    // response can be "{\"status\": \"idle\"}"
+    public static func asyncOpen(jobId: UInt64) async -> String {
+        debugPrint("\(#function) starts")
+        let jsonStr = String(ZanoCore.try_pull_result(jobId))
+        debugPrint("jsonStr")
+        debugPrint(jsonStr)
+        debugPrint("=======")
         
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseJsonResponse(jsonData: jsonData, resultType: WalletResult.self)
-        return result
+        return jsonStr
     }
 
-    // required to get JobId from asyncCall
-    public static func asyncClose(jobId: UInt64) async throws -> SuccessResult {
+    public static func asyncClose(jobId: UInt64) async throws -> String {
+        debugPrint("\(#function) starts")
         let jsonStr = String(String(ZanoCore.try_pull_result(jobId)))
-        
-        guard let jsonData = jsonStr.data(using: .utf8) else {
-            throw ZANOError.conversionFailure(message: "")
-        }
-        
-        let result = try JSONRPCParser.parseJsonResponse(jsonData: jsonData, resultType: SuccessResult.self)
-        return result
+        return jsonStr
     }
     
     // Currently ZANO always retuns 10000000000(0.01)
@@ -263,48 +215,10 @@ public enum ZanoWallet {
     }
 }
 
-public struct WorkingDir {
-    public static var dir: String? {
-        let fileManager = FileManager.default
-
-        if let docsDir = try? fileManager.url(for: .documentDirectory, in: .userDomainMask, appropriateFor: nil, create: true) {
-            return docsDir.path
-        } else {
-            return nil
-        }
-    }
-}
-
-public struct ZANOConst {
-    public static let ZANO_TESTNET_IP = "195.201.107.230"
-    public static let ZANO_TESTNET_PORT = "33336"
-    public static let TEST_WALLET_NAME = "testWallet1"
-    public static let TEST_WALLET_PASSWORD = "testWalletPassword"
-    public static let TEST_WALLET_SEED =
-        "insult understand back vein peaceful somewhere bare deeply anyone hunger wait prince shoot grip scream bowl mama card dread less sadness angry explore almost worry dim"
-    public static let TEST_WALLET_ADDRESS =
-        "ZxCx5QqAh67JwX7F2hs7M2FcJ3bcLEhWfb2yqiofDeWn67guYBA4fwb1ekbsp7vZid35fVJtY6E2iiWVUJWRkyQu1Mchqr3iE"
-}
-
-// MARK: = Prser
-
-public struct JSONRPCParser {
-    public static func parseResponse<T: Codable>(jsonData: Data, resultType: T.Type) throws -> T {
-        let decodedResponse = try JSONDecoder().decode(RPCResponse.self, from: jsonData)
-        
-        if let result = decodedResponse.result {
-            // Return success with the parsed result
-            return try result.get(T.self)
-        } else if let error = decodedResponse.error {
-            // Return failure with the error
-            throw ZANOError.errorResponse(code: error.code, message: error.message)
-        } else {
-            throw ZANOError.unknown(message: "")
-        }
-    }
-    
-    public static func parseJsonResponse<T: Codable>(jsonData: Data, resultType: T.Type) throws -> T {
-        return try JSONDecoder().decode(T.self, from: jsonData)
-    }
+public enum ZANOError: Error {
+    case errorResponse(code: String, message: String)
+    case asyncErrorResponse(status: String, code: String, message: String)
+    case unknown(message: String)
+    case conversionFailure(message: String)
 }
