+++ zano-iosTests/zano_iosTests.swift
@@ -99,28 +99,74 @@ class ParseTests: XCTestCase {
         XCTAssertEqual(walletResponse.walletInfo.viewSecKey, "7149d1f592ed7e03719e7451a23ebf2250016384f4d7cf057241544648fe4d01")
     }
     
-    func testAsyncResponseSuccess() throws {
-        let encoded = mockAsyncResponse.data(using: .utf8)!
-        let asyncResult = try JSONRPCParser.parseJsonResponse(jsonData: encoded, resultType: AsyncResult.self)
-        let syncSuccessResponse = try AsyncSuccessResponse.create(asyncResult: asyncResult)
-        XCTAssertEqual(syncSuccessResponse.status, .delivered)
-        XCTAssertEqual(syncSuccessResponse.balance.balance, 4990000000000)
-        XCTAssertEqual(syncSuccessResponse.balance.balances.count, 1)
-        let balance = syncSuccessResponse.balance.balances.first!
-        XCTAssertEqual(balance.awaitingIn, 0)
-        XCTAssertEqual(balance.awaitingOut, 0)
-        XCTAssertEqual(balance.total, 5000000000000)
-        XCTAssertEqual(balance.unlocked, 4990000000000)
-        XCTAssertEqual(balance.assetInfo.currentSupply, 0)
-        XCTAssertEqual(balance.assetInfo.decimalPoint, 12)
-        XCTAssertEqual(balance.assetInfo.fullName, "Zano")
-        XCTAssertFalse(balance.assetInfo.hiddenSupply)
-        XCTAssertEqual(balance.assetInfo.metaInfo, "")
-        XCTAssertEqual(balance.assetInfo.owner, "0000000000000000000000000000000000000000000000000000000000000000")
-        XCTAssertEqual(balance.assetInfo.ticker, "ZANO")
-        XCTAssertEqual(balance.assetInfo.totalMaxSupply, 0)
-        XCTAssertEqual(syncSuccessResponse.balance.unlocked_balance, 4990000000000)
+    func testAsyncResponseIdle() throws {
+        let encoded = mockAsyncResponseIdle.data(using: .utf8)!
+        let asyncResult = try JSONRPCParser.parseJsonResponse(jsonData: encoded, resultType: AsyncOpenSuccessResponse.self)
+        XCTAssertEqual(asyncResult.status, .idle)
     }
+
+    
+//    func testAsyncResponseSuccess() throws {
+//        let encoded = mockAsyncResponse.data(using: .utf8)!
+//        let syncSuccessResponse = try JSONRPCParser.parseJsonResponse(jsonData: encoded, resultType: AsyncOpenSuccessResponse.self)
+////        let syncSuccessResponse = try AsyncSuccessResponse.create(asyncResult: asyncResult)
+//        XCTAssertEqual(syncSuccessResponse.status, .delivered)
+//        XCTAssertEqual(syncSuccessResponse.walletResult.balance, 4990000000000)
+//        XCTAssertEqual(syncSuccessResponse.balance?.balances.count, 1)
+//        let balance = syncSuccessResponse.balance!.balances.first!
+//        XCTAssertEqual(balance.awaitingIn, 0)
+//        XCTAssertEqual(balance.awaitingOut, 0)
+//        XCTAssertEqual(balance.total, 5000000000000)
+//        XCTAssertEqual(balance.unlocked, 4990000000000)
+//        XCTAssertEqual(balance.assetInfo.currentSupply, 0)
+//        XCTAssertEqual(balance.assetInfo.decimalPoint, 12)
+//        XCTAssertEqual(balance.assetInfo.fullName, "Zano")
+//        XCTAssertFalse(balance.assetInfo.hiddenSupply)
+//        XCTAssertEqual(balance.assetInfo.metaInfo, "")
+//        XCTAssertEqual(balance.assetInfo.owner, "0000000000000000000000000000000000000000000000000000000000000000")
+//        XCTAssertEqual(balance.assetInfo.ticker, "ZANO")
+//        XCTAssertEqual(balance.assetInfo.totalMaxSupply, 0)
+//        XCTAssertEqual(syncSuccessResponse.balance?.unlocked_balance, 4990000000000)
+//    }
+//    
+    func testAsyncOpenResponseSuccess() throws {
+        let encoded = mockAsyncOpenWalletSuccessResponse.data(using: .utf8)!
+        let asyncResult = try JSONRPCParser.parseJsonResponse(jsonData: encoded, resultType: AsyncResult.self)
+        let syncSuccessResponse = try AsyncOpenSuccessResponse.create(asyncResult: asyncResult)
+        XCTAssertEqual(syncSuccessResponse.status, .delivered)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.name, "")
+        XCTAssertEqual(syncSuccessResponse.walletResult?.pass, "")
+        XCTAssertFalse(syncSuccessResponse.walletResult?.recovered ?? true)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.seed, "kitchen patience fully might bit worthless lead early language ghost blue look heel outside crumble creation everytime replace painful hidden underneath grew understand horizon war replace")
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletFileSize, 528)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletID, 0)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletLocalBCSize, 1)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.address, "ZxCW6Gtijrh5Z5onFNW66hVZoBi5P3X4XJ5rBqafTMuUgoUrvu6XxYSHdg3VWniVf97kHozLKj7MgcA9PdwJPZNL1SwivnU7t")
+        XCTAssertEqual(syncSuccessResponse.walletResult?.recentHistory.lastItemIndex, 0)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.recentHistory.totalHistoryItems, 0)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.count, 1)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.first!.assetInfo.assetID, "d6329b5b1f7c0805b5c345f4957554002a2f557845f64d7645dae0e051a6498a")
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.first!.assetInfo.currentSupply, 0)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.first!.assetInfo.decimalPoint, 12)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.first!.assetInfo.fullName, "Zano")
+        XCTAssertFalse(syncSuccessResponse.walletResult?.walletInfo.balances.first!.assetInfo.hiddenSupply ?? true)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.first!.assetInfo.currentSupply, 0)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.first!.assetInfo.metaInfo, "")
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.first!.assetInfo.ticker, "ZANO")
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.first!.assetInfo.totalMaxSupply, 0)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.first!.awaitingIn, 0)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.first!.awaitingOut, 0)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.first!.total, 0)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.balances.first!.unlocked, 0)
+        XCTAssertFalse(syncSuccessResponse.walletResult?.walletInfo.hasBareUnspentOutputs ?? true)
+        XCTAssertFalse(syncSuccessResponse.walletResult?.walletInfo.isAuditable ?? true)
+        XCTAssertFalse(syncSuccessResponse.walletResult?.walletInfo.isWatchOnly ?? true)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.minedTotal, 0)
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.path, "/Users/user/Library/Developer/CoreSimulator/Devices/11C27930-183C-4F19-BDDB-0AF90C77BC14/data/Containers/Data/Application/8A192E13-38E8-462C-B5C0-3D262552F8BC/Documents/wallets/testWallet1")
+        XCTAssertEqual(syncSuccessResponse.walletResult?.walletInfo.viewSecKey, "7149d1f592ed7e03719e7451a23ebf2250016384f4d7cf057241544648fe4d01")
+    }
+
+
     func testWalletStatus() throws {
         let encoded = mockWalletStatus.data(using: .utf8)!
         let result = try JSONRPCParser.parseJsonResponse(jsonData: encoded, resultType: WalletSyncStatus.self)
@@ -143,6 +189,12 @@ class ParseTests: XCTestCase {
           "progress": 0,
           "wallet_state": 1
         }
+        """
+    }
+    
+    var mockAsyncResponseIdle: String {
+        return """
+         {"status": "idle"}
         """
     }
         
@@ -175,6 +227,60 @@ class ParseTests: XCTestCase {
       }  }
  """
     }
+    
+    var mockAsyncOpenWalletSuccessResponse: String {
+        return """
+ {
+   "status": "delivered",
+   "result": {
+     "id": 0,
+     "jsonrpc": "",
+     "result": {
+       "name": "",
+       "pass": "",
+       "recent_history": {
+         "last_item_index": 0,
+         "total_history_items": 0
+       },
+       "recovered": false,
+       "seed": "kitchen patience fully might bit worthless lead early language ghost blue look heel outside crumble creation everytime replace painful hidden underneath grew understand horizon war replace",
+       "wallet_file_size": 528,
+       "wallet_id": 0,
+       "wallet_local_bc_size": 1,
+       "wi": {
+         "address": "ZxCW6Gtijrh5Z5onFNW66hVZoBi5P3X4XJ5rBqafTMuUgoUrvu6XxYSHdg3VWniVf97kHozLKj7MgcA9PdwJPZNL1SwivnU7t",
+         "balances": [
+           {
+             "asset_info": {
+               "asset_id": "d6329b5b1f7c0805b5c345f4957554002a2f557845f64d7645dae0e051a6498a",
+               "current_supply": 0,
+               "decimal_point": 12,
+               "full_name": "Zano",
+               "hidden_supply": false,
+               "meta_info": "",
+               "owner": "0000000000000000000000000000000000000000000000000000000000000000",
+               "ticker": "ZANO",
+               "total_max_supply": 0
+             },
+             "awaiting_in": 0,
+             "awaiting_out": 0,
+             "total": 0,
+             "unlocked": 0
+           }
+         ],
+         "has_bare_unspent_outputs": false,
+         "is_auditable": false,
+         "is_watch_only": false,
+         "mined_total": 0,
+         "path": "/Users/user/Library/Developer/CoreSimulator/Devices/11C27930-183C-4F19-BDDB-0AF90C77BC14/data/Containers/Data/Application/8A192E13-38E8-462C-B5C0-3D262552F8BC/Documents/wallets/testWallet1",
+         "view_sec_key": "7149d1f592ed7e03719e7451a23ebf2250016384f4d7cf057241544648fe4d01"
+       }
+     }
+   }
+ }
+ """
+    }
+
     
     var mockConnectivityStatus: String {
         return """
